Refactor code into modules and update docs
